+++ src/test/java/antigravity/service/ProductResponseServiceTest.java
@@ -7,23 +7,15 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.BDDMockito.then;
 
 @DisplayName("서비스 - 제품 요청 서비스 테스트")
 @ExtendWith(MockitoExtension.class)
-class ProductRequestServiceTest {
+class ProductResponseServiceTest {
 
     @InjectMocks
-    private ProductRequestService sut;
+    private ProductResponseService sut;
     @Mock
     private ProductRepository productRepository;
