+++ src/test/java/antigravity/service/ProductServiceTest.java
@@ -1,23 +1,18 @@
 package antigravity.service;
 
-import antigravity.entity.Product;
 import antigravity.repository.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @DisplayName("서비스 - 제품 서비스 테스트")
 @ExtendWith(MockitoExtension.class)
 class ProductServiceTest {
 
     @InjectMocks
-    private ProductService sut;
+    private ProductRequestService sut;
     @Mock
     private ProductRepository eventRepository;
 }
